requirements.txt
streamlit
cohere
PyMuPDF
pdf2image
pytesseract
Pillow
python-docx
python-pptx
fpdf
